Add payment types for testing and fix specs
