+++ spec/support/payments/default.rb
@@ -1,0 +1,7 @@
+module Payments
+  class Default < Base
+    def payment_type
+      "default"
+    end
+  end
+end
