+++ spec/support/payments/master.rb
@@ -1,0 +1,7 @@
+module Payments
+  class Master < Base
+    def payment_type
+      "master"
+    end
+  end
+end
