+++ spec/support/payments/recurring.rb
@@ -1,0 +1,7 @@
+module Payments
+  class Recurring < Base
+    def payment_type
+      "recurring"
+    end
+  end
+end
